+++ .storybook/blocks/index.ts
@@ -3,3 +3,4 @@ export { CssVars } from './css-vars';
 export { ComponentHeader } from './component-header';
 export { Row, Col, Container } from './grid';
 export { Tabs } from './tabs';
+export { ArgsTabs } from './args-tabs';