+++ packages/input/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## 12.1.4
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   shared@0.5.0
+    -   button@9.0.5
+
 ## 12.1.3
 
 ### Patch Changes