+++ packages/intl-phone-input/CHANGELOG.md
@@ -1,5 +1,21 @@
 # Change Log
 
+## 9.5.11
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   input-autocomplete@10.1.6
+    -   select@15.1.6
+
+## 9.5.10
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   select@15.1.5
+    -   input-autocomplete@10.1.5
+
 ## 9.5.9
 
 ### Patch Changes