+++ packages/picker-button/CHANGELOG.md
@@ -1,5 +1,20 @@
 # Change Log
 
+## 11.1.6
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   button@9.0.5
+    -   select@15.1.6
+
+## 11.1.5
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   select@15.1.5
+
 ## 11.1.4
 
 ### Patch Changes