+++ packages/with-suffix/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## 4.1.7
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   input@12.1.4
+
 ## 4.1.6
 
 ### Patch Changes